requirements.txt
asttokens==2.0.8
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
certifi @ file:///C:/Windows/TEMP/abs_e9b7158a-aa56-4a5b-87b6-c00d295b01fanefpc8_o/croots/recipe/certifi_1655968940823/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.0.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
ipykernel==6.15.2
ipython==8.5.0
jedi==0.18.1
jupyter-core==4.11.1
jupyter_client==7.3.5
matplotlib-inline==0.1.6
MouseInfo==0.1.3
nest-asyncio==1.5.5
outcome==1.2.0
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
playwright==1.25.2
prompt-toolkit==3.0.31
psutil==5.9.2
pure-eval==0.2.2
PyAutoGUI==0.9.53
pycparser==2.21
pyee==8.1.0
PyGetWindow==0.0.9
Pygments==2.13.0
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytweening==1.0.4
pywin32==304
pyzmq==23.2.1
requests==2.28.1
selenium==4.4.3
selenium-stealth==1.0.6
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
stack-data==0.5.0
TikTokApi==5.2.2
tornado==6.2
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
wcwidth==0.2.5
websockets==10.1
wincertstore==0.2
wsproto==1.2.0
